requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
Flask-SocketIO==5.5.1
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.3.0
gunicorn==21.2.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.14.3
simple-websocket==1.1.0
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.3.1